# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Hub API",
        "description": "API para gerenciamento de eventos com venda de ingressos, processamento de pagamentos e notificações.",
        "contact": {
            "email": "contato@eventhub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/public/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Criar usuário público (Participante)",
                "description": "Cria um novo usuário participante sem autenticação",
                "operationId": "b9e7790a71f088c9a5b38c944ba33dc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "João Silva"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "joao@example.com"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "PARTICIPANT"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    }
                }
            }
        },
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Criar usuário (Admin)",
                "description": "Cria um novo usuário (Participante ou Organizador). Requer autenticação de Admin para criar Organizador",
                "operationId": "019c4bcb7000fd1fb597d07b8caec124",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Santos"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "PARTICIPANT",
                                            "ORGANIZER"
                                        ],
                                        "example": "ORGANIZER"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Maria Santos"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "maria@example.com"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "ORGANIZER"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via token Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}